% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crop_par.R
\name{get_crop_par}
\alias{get_crop_par}
\title{get_crop_par}
\usage{
get_crop_par(crop, model_dir)
}
\arguments{
\item{crop}{Nombre del cultivo (4 letras, e.g. "SOYB")}

\item{model_dir}{Directorio del proyecto SWAT}
}
\description{
Obtiene los parámetros de un cultivo
}
\examples{
#get_crop_par(crop="SOYB",model_dir = "C:/model")
}
