% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSWATmodel.R
\name{installSWATmodel}
\alias{installSWATmodel}
\title{installSWATmodel}
\usage{
installSWATmodel(model_dir, model_version)
}
\arguments{
\item{model_dir}{Directorio en donde descomprimir los archivos TxTinout}

\item{model_version}{Versión del modelo a instalar (e.g. "TALA_20210419")}
}
\description{
Descomprime una versión del modelo en el directorio "model_dir"
}
