% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TalaToolsSWAT-package.R
\docType{package}
\name{TalaToolsSWAT-package}
\alias{TalaToolsSWAT}
\alias{TalaToolsSWAT-package}
\title{TalaToolsSWAT: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
